requirements.txt
streamlit>=1.29.0
joblib>=1.3.0
pandas>=2.1.0
geopy>=2.4.0
requests>=2.31.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0